backend/requirements.txt
alembic==1.16.4
asyncpg==0.30.0
aiofiles==24.1.0
email-validator==2.2.0
fastapi==0.116.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
redis==6.4.0
SQLAlchemy==2.0.43
structlog==25.4.0
uvicorn[standard]==0.35.0
